apps/api/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
pillow 
pillow-heif 
python-magic